[test refactor] Continue test refactoring from vows to mocha.
