+++ UPGRADING.md
@@ -1,0 +1,5 @@
+
+### Braindump of breaking changes
+
+- `winston.Logger` will no longer respond with an error when logging with no transports
+- `winston.Logger` will no longer respond with an error if the same transports are added twice.
