+++ lib/winston/log-stream.js
@@ -77,7 +77,6 @@ LogStream.prototype.setLevels = function (target) {
   return this;
 };
 
-
 /*
  * @private function _transform (obj)
  * Pushes data so that it can be picked up by all of
